+++ src/cli.rs
@@ -13,7 +13,7 @@ pub struct Cli {
 #[derive(Debug, Subcommand)]
 pub enum Commands {
     #[command(about = "display configuration settings")]
-    Cfg,
+    Setup,
     #[command(about = "downloads the IB Flex Cash Tx report")]
     Dl(DlParams),
     #[command(about = "compares IB Flex Cash Tx report and Ledger")]