+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - aligning amounts in the comparison display
+- the comparison transactions now also sorted by type
 
 ### Changed