+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2023-02-21
+
+### Changed
+
+- changed Config command to Setup, to free the "c" shortcut for Compare
+
 ### Security
 
 - updated dependencies